Organised graphql_core tests.
